+++ src/main/kotlin/ru/pochtabank/templates/core/mvvm/src/app_package/blankUiState.kt
@@ -1,17 +1,17 @@
 package ru.pochtabank.templates.viewModelFragment.src.app_package
 
 import com.android.tools.idea.wizard.template.escapeKotlinIdentifier
+import ru.pochtabank.templates.models.BuildSettings
 
 fun blankUiStateKt(
     className: String,
     packageName: String,
+    buildSettings: BuildSettings,
 ): String {
     return """package ${escapeKotlinIdentifier(packageName)}
 
-internal sealed class $className {
+${buildSettings.visibilityType.getClassVisibility()}sealed class $className {
     //Реализацию писать внутри фигурных скобок! 
-    // ${packageName}
-    // ${'$'}{escapeKotlinIdentifier(packageName)
 }
 """
 }