+++ src/main/kotlin/ru/pochtabank/templates/core/utils/Constants.kt
@@ -1,0 +1,5 @@
+package ru.pochtabank.templates.core.utils
+
+object Constants {
+    const val MIN_SDK = 21
+}