+++ src/main/kotlin/ru/pochtabank/templates/models/BuildSettings.kt
@@ -1,6 +1,7 @@
-package ru.pochtabank.templates.viewModelFragment
+package ru.pochtabank.templates.models
 
 import com.android.tools.idea.wizard.template.ModuleTemplateData
+import ru.pochtabank.templates.models.VisibilityType
 
 class BuildSettings(
     val moduleData: ModuleTemplateData,
@@ -9,4 +10,5 @@ class BuildSettings(
     val layoutName: String,
     val isGenerateUiStateModel: Boolean,
     val isGenerateRoutingStateModel: Boolean,
+    val visibilityType: VisibilityType,
 )