+++ src/main/kotlin/ru/pochtabank/templates/newModuleModule/emptyActivityKt.kt
@@ -1,0 +1,33 @@
+package ru.pochtabank.templates.newModuleModule
+
+import com.android.tools.idea.wizard.template.escapeKotlinIdentifier
+import com.android.tools.idea.wizard.template.getMaterialComponentName
+import com.android.tools.idea.wizard.template.renderIf
+
+fun emptyActivityKt(
+    packageName: String, namespace: String, activityClass: String, layoutName: String, generateLayout: Boolean, useAndroidX: Boolean
+) = """
+package ${escapeKotlinIdentifier(packageName)}
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import ${getMaterialComponentName("android.support.v7.app.AppCompatActivity", useAndroidX)}
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+${renderIf(namespace != packageName) { "import ${escapeKotlinIdentifier(namespace)}.R" }}
+
+class $activityClass : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        ${renderIf(generateLayout) {
+    """enableEdgeToEdge()
+        setContentView(R.layout.$layoutName)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
+            insets
+        }"""
+}}
+    }
+}
+"""