+++ src/main/kotlin/ru/pochtabank/templates/viewModelFragment/viewModelFragmentRecipe.kt
@@ -3,14 +3,19 @@ package ru.pochtabank.templates.viewModelFragment
 import com.android.tools.idea.wizard.template.RecipeExecutor
 import com.android.tools.idea.wizard.template.impl.activities.common.addAllKotlinDependencies
 import ru.pochtabank.templates.core.mvvm.src.app_package.blankViewModelKt
+import ru.pochtabank.templates.core.utils.getModuleName
 import ru.pochtabank.templates.core.utils.getRouteStateModelName
 import ru.pochtabank.templates.core.utils.getUiModelsPackageName
 import ru.pochtabank.templates.core.utils.getUiStateModelName
+import ru.pochtabank.templates.core.utils.presentationPackageEndWith
 import ru.pochtabank.templates.core.utils.uiModelsPathName
+import ru.pochtabank.templates.models.BuildSettings
+import ru.pochtabank.templates.models.VisibilityType
 import ru.pochtabank.templates.viewModelFragment.res.layout.blankFragmentXml
 import ru.pochtabank.templates.viewModelFragment.src.app_package.blankFragmentKt
 import ru.pochtabank.templates.viewModelFragment.src.app_package.blankRoutingStateKt
 import ru.pochtabank.templates.viewModelFragment.src.app_package.blankUiStateKt
+import ru.pochtabank.templates.viewModelFragment.src.app_package.buildScreenFactoryKt
 
 fun RecipeExecutor.viewModelFragmentRecipe(
     buildSettings: BuildSettings,
@@ -41,6 +46,7 @@ fun RecipeExecutor.viewModelFragmentRecipe(
             blankUiStateKt(
                 className = getUiStateModelName(buildSettings.entityName),
                 packageName = getUiModelsPackageName(buildSettings.packageName),
+                buildSettings = buildSettings,
             ),
             srcOut.resolve("$uiModelsPathName/${getUiStateModelName(buildSettings.entityName)}.kt")
         )
@@ -50,10 +56,28 @@ fun RecipeExecutor.viewModelFragmentRecipe(
             blankRoutingStateKt(
                 className = getRouteStateModelName(buildSettings.entityName),
                 packageName = getUiModelsPackageName(buildSettings.packageName),
+                buildSettings = buildSettings,
             ),
             srcOut.resolve("$uiModelsPathName/${getRouteStateModelName(buildSettings.entityName)}.kt")
         )
     }
+    if (buildSettings.visibilityType == VisibilityType.PUBLIC_FACTORY) {
+        val path = if (buildSettings.packageName.contains(presentationPackageEndWith)) {
+            srcOut
+                .resolveSibling("")
+                .resolveSibling("${getModuleName(buildSettings.moduleData).replaceFirstChar(Char::uppercase)}ScreenFactory.kt")
+        } else {
+            srcOut.resolve("${getModuleName(buildSettings.moduleData).replaceFirstChar(Char::uppercase)}ScreenFactory.kt")
+        }
+
+        save(
+            buildScreenFactoryKt(
+                buildSettings = buildSettings
+            ),
+            path
+        )
+    }
+
     save(
         blankFragmentXml(),
         resOut.resolve("layout/${buildSettings.layoutName}.xml")