+++ CHANGELOG.md
@@ -3,5 +3,5 @@
 # PochtaBankMvvmTemplate Changelog
 
 ## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+### Добавлено
+- Возможность создания Fragment + ViewModel (+ вспомогательные классы);